+++ src/lib/components/icon/types.ts
@@ -1,0 +1,3 @@
+import type { SvgIconKeys } from "./SvgIcon";
+
+export type IconKeys = SvgIconKeys | "circle";